rename: 컴포넌트 이름 파스칼 케이스로 변경
